Various improvments from beta testing with Fabien
Add MSVC runtime back in. This was previously automatically
taken care of inside of the Postgres install. (Thanks to Fabien
for finding a test environment to incover this problem.)

Do not try to wipe out the data directory, that's too dangerous.
(Why did it ever work like this?)

Remove more unused code (simplify /PGINSTDIR setting).

Tune Postgres correctly based on the size of the servers we use.