A comment on VC redistributables
There are two VC runtimes checked in here. One is
to make the building tools work correctly on the build
server (lib/vcredist_x86.exe).

The other one (windows/vcredist_x86.exe) is the one that
is included in the final AIO, which is needed so that
PostgreSQL will run correctly on the Unifield servers.